refactor(ui): standardize cn utility import across components
Update all UI components to import cn utility from @/lib/utils instead
of @/utils/ui for consistency with shadcn/ui conventions.

- Update badge.jsx to use @/lib/utils
- Update button.jsx to use @/lib/utils
- Update card.jsx to use @/lib/utils
- Update progress.jsx to use @/lib/utils
- Update select.jsx to use @/lib/utils
- Update separator.jsx to use @/lib/utils
- Update tabs.jsx to use @/lib/utils

This ensures all components follow the same import pattern and use the
canonical cn utility function from lib/utils.js.
