+++ src/components/ui/card.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { cn } from "../../utils/ui";
+import { cn } from "../../lib/utils";
 
 const Card = React.forwardRef(({ className, ...props }, ref) => (
   <div
