+++ src/components/ui/select.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { cn } from "../../utils/ui";
+import { cn } from "../../lib/utils";
 
 const Select = React.forwardRef(({ className, ...props }, ref) => (
   <select
