+++ src/components/ui/tabs.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { cn } from "../../utils/ui";
+import { cn } from "../../lib/utils";
 
 const Tabs = React.forwardRef(({ className, ...props }, ref) => (
   <div ref={ref} className={cn("", className)} {...props} />
